🚚 rename:  컴포넌트와 스타일드 속성 분리 (#18)
